requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rich>=13.7.0
typer>=0.12.0
tavily-python>=0.3.0
reportlab>=4.0.0

------------
sentence-transformers>=2.7.0
apscheduler>=3.10.0
pdfplumber>=0.10.0
scikit-learn>=1.4.0
numpy>=1.26.0